// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/public-recipe-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PublicRecipeStepHttpResponse } from './public-recipe-step-http-response';

export interface PublicRecipeHttpResponse {
    authorName?: string | null;
    category?: string | null;
    cookTime?: number | null;
    description?: string | null;
    id?: string;
    imageUrl?: string | null;
    images?: Array<string>;
    language?: string;
    missingIngredientCount?: number;
    missingIngredientNames?: Array<string>;
    name?: string;
    prepTime?: number | null;
    servings?: number;
    steps?: Array<PublicRecipeStepHttpResponse>;
    totalAlcohol?: number | null;
    totalCalories?: number | null;
    totalCarbs?: number | null;
    totalFats?: number | null;
    totalFiber?: number | null;
    totalProteins?: number | null;
}